Refactoring forever.
