+++ src/libcrow/io.hpp
@@ -1,7 +1,7 @@
 #ifndef CROW_STDLIB_IO_HPP
 #define CROW_STDLIB_IO_HPP
 
-namespace stdlib {
+namespace libcrow {
 template<typename... Args>
 inline auto print(std::format_string<Args...> t_fmt, Args&&... t_args) -> void
 {
@@ -13,6 +13,6 @@ inline auto println(std::format_string<Args...> t_fmt, Args&&... t_args) -> void
 {
   std::println(fmt, std::forward<Args>(t_args)...);
 }
-} // namespace stdlib
+} // namespace libcrow
 
 #endif // CROW_STDLIB_IO_HPP
